+++ src/index.ts
@@ -1,5 +1,3 @@
-export { getAPI as getCalendarAPI } from './lib/api/calendar';
-export { getAPI as getYoutubeAPI } from './lib/api/youtube';
-export { getItems } from './lib/api/shared';
+export { getApi } from './lib/api';
 export { createProfile, getProfiles } from './lib/profiles';
 export { login, getAuth } from './lib/auth';